+++ utility/content/isPostLinkGeneratedQuote.js
@@ -1,3 +1,5 @@
+import isPostLinkQuote from './isPostLinkQuote.js'
+
 export default function isPostLinkGeneratedQuote(postLink) {
-	return Array.isArray(postLink.content) && postLink.content[0].type === 'quote' && postLink.content[0].generated
+	return isPostLinkQuote(postLink) && postLink.content[0].contentGenerated
 }